Create PreAPITeleportEvent
As spigot api teleportation is subject to stricter checks than the
default minecraft teleportation, some plugins may not be able to
teleport entities that would be perfectly fine teleporting under the
vanilla standart.

To not remove the created rules by spigot and potentially break plugin
compatibility the PreAPITeleportEvent is called prior to most checks
executed by spigot which allows plugins to modify potential target
entities that are teleported by other plugins prior to the checks.